refactor: split analysis/subprocess.py into focused modules
